Move print_settlements into DriverLogger
