+++ crates/solver/src/driver_logger.rs
@@ -13,11 +13,12 @@ use contracts::GPv2Settlement;
 use gas_estimation::GasPrice1559;
 use itertools::Itertools;
 use model::order::{Order, OrderKind};
+use num::{BigRational, ToPrimitive};
 use primitive_types::H256;
 use shared::Web3;
 use std::sync::Arc;
 use tracing::{Instrument as _, Span};
-use web3::types::TransactionReceipt;
+use web3::types::{AccessList, TransactionReceipt};
 
 pub struct DriverLogger {
     pub metrics: Arc<dyn SolverMetrics>,
@@ -184,5 +185,81 @@ impl DriverLogger {
             }
         };
         tokio::task::spawn(task.instrument(Span::current()));
+    }
+
+    pub fn print_settlements(
+        rated_settlements: &[(Arc<dyn Solver>, RatedSettlement, Option<AccessList>)],
+        fee_objective_scaling_factor: &BigRational,
+    ) {
+        let mut text = String::new();
+        for (solver, settlement, access_list) in rated_settlements {
+            use std::fmt::Write;
+            write!(
+                text,
+                "\nid={} solver={} \
+             objective={:.2e} surplus={:.2e} \
+             gas_estimate={:.2e} gas_price={:.2e} \
+             unscaled_unsubsidized_fee={:.2e} unscaled_subsidized_fee={:.2e} \
+             access_list_addreses={}",
+                settlement.id,
+                solver.name(),
+                settlement.objective_value().to_f64().unwrap_or(f64::NAN),
+                settlement.surplus.to_f64().unwrap_or(f64::NAN),
+                settlement.gas_estimate.to_f64_lossy(),
+                settlement.gas_price.to_f64().unwrap_or(f64::NAN),
+                (&settlement.scaled_unsubsidized_fee / fee_objective_scaling_factor)
+                    .to_f64()
+                    .unwrap_or(f64::NAN),
+                settlement
+                    .unscaled_subsidized_fee
+                    .to_f64()
+                    .unwrap_or(f64::NAN),
+                access_list.clone().unwrap_or_default().len()
+            )
+            .unwrap();
+        }
+        tracing::info!("Rated Settlements: {}", text);
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::solver::dummy_arc_solver;
+
+    #[test]
+    #[ignore]
+    fn print_settlements() {
+        let a = [
+            (
+                dummy_arc_solver(),
+                RatedSettlement {
+                    id: 0,
+                    settlement: Default::default(),
+                    surplus: BigRational::new(1u8.into(), 1u8.into()),
+                    unscaled_subsidized_fee: BigRational::new(2u8.into(), 1u8.into()),
+                    scaled_unsubsidized_fee: BigRational::new(3u8.into(), 1u8.into()),
+                    gas_estimate: 4.into(),
+                    gas_price: BigRational::new(5u8.into(), 1u8.into()),
+                },
+                None,
+            ),
+            (
+                dummy_arc_solver(),
+                RatedSettlement {
+                    id: 6,
+                    settlement: Default::default(),
+                    surplus: BigRational::new(7u8.into(), 1u8.into()),
+                    unscaled_subsidized_fee: BigRational::new(8u8.into(), 1u8.into()),
+                    scaled_unsubsidized_fee: BigRational::new(9u8.into(), 1u8.into()),
+                    gas_estimate: 10.into(),
+                    gas_price: BigRational::new(11u8.into(), 1u8.into()),
+                },
+                None,
+            ),
+        ];
+
+        shared::tracing::initialize_for_tests("INFO");
+        DriverLogger::print_settlements(&a, &BigRational::new(1u8.into(), 2u8.into()));
     }
 }
